Refactor store into multiple files
