+++ lib/bot.js
@@ -11,7 +11,6 @@ class Bot {
   myInfo: any
   chat: any
 
-  // Create a new store for each bot initialized
   constructor() {
     const _store = createStore(reducer, initialState)
     const base = Base(_store)
