+++ lib/store/actions.js
@@ -1,0 +1,23 @@
+// @flow
+import type {Action} from './types'
+
+/*
+ * Actions
+ */
+type Props = {|
+  username: string,
+  devicename: string,
+  flags: {|
+    verbose: boolean,
+  |},
+|}
+
+export const actionInit = (props: Props): Action => ({
+  type: 'INIT',
+  payload: props,
+})
+
+export const actionDeinit = (): Action => ({
+  type: 'DEINIT',
+  payload: {},
+})
